add dynopt unit tests

+++ src/HeuristicLib/Problems/Dynamic/EvaluationTiming.cs
@@ -1,3 +1,3 @@
 ﻿namespace HEAL.HeuristicLib.Problems.Dynamic;
 
-public readonly record struct EvaluationTiming(int EpochCount, int Epoch, bool Valid);
+public readonly record struct EvaluationTiming(long EpochCount, int Epoch, bool Valid);
